Implement new booking page, service requests attachments, project restructure and unit, integration, concurrent testing

+++ user-service/src/main/java/com/hse/userservice/internal/dto/InternalServiceRequestDetailDto.java
@@ -10,7 +10,7 @@ public record InternalServiceRequestDetailDto(
         String userEmail,
         String typeName,
         String name,
-        Integer cost,
+        Long cost,
         Long balanceMinorUnits,
         String status,
         LocalDateTime createdAt,
